+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Plugins
    - Proxy -- Fixed typo in code which prevented headers from being properly
    forwarded which results in non-existent content-types which prevented proper parsing. [Issue #135]
+- Modules
+   - Unencrypted password forms -- Checks for non-nil form fields before iterating. [Issue #136]
 
 ## Version 0.4.0.1 _(January 9, 2012)_
 - Reports