+++ CHANGELOG.md
@@ -6,6 +6,8 @@
    - Proxy -- Fixed typo in code which prevented headers from being properly.
    forwarded which results in non-existent content-types which prevented proper parsing. [Issue #135]
 - Modules
+   - API
+      - Updated to provide access to running plugins.
    - Unencrypted password forms -- Checks for non-nil form fields before iterating. [Issue #136]
 - Extras
    - All modules under <em>/extras</em> had to be removed because they distributed GPLv3 licensed content.